# Optimize grammar generator