This part of the code instantiates parametric productions with real sorts for the parser.

The old code was `O(no_sorts^no_params)` for each production in a module. It would create that production and then see if it already exists.
The new code simply creates only the productions that are needed in the end. Complexity comes down to linear.

Here is some detailed timing info from an extreme example (KEVM and C below):
Definition: `mkr-mcd-spec`
cmd: `time kompile -v --backend llvm kmcd-prelude.md ...`
Before:
```
...
Parse configurations [24/24 declarations]                    = 01m 04s
Parse rules [1728/1728 rules]                                = 02m 56s
Total                                                        = 04m 59s
```
After: 
```
...
Parse configurations [24/24 declarations]                    =  4.820s
Parse rules [1728/1728 rules]                                = 02m 33s
Total                                                        = 03m 29s
```
When we would have lots and lots of sorts, the time spent in this part of the code was quite bad. Here are some examples with before and after timings for specific modules (ms):
```
Module                 before -> after
JOIN-CONFIG-CELLS:       1795 ->  8
JUG-CONFIG-CELLS:        1460 ->  7
KMCD-CONFIG-CELLS:       7828 -> 17
KMCD-PROPS-CONFIG-CELLS: 8276 -> 32
KMCD-GEN-CONFIG-CELLS:   8639 -> 18
...
KMCD-RULE-CELLS:         25731 -> 208
KMCD-PROPS-RULE-CELLS:   26043 -> 203
KMCD-PRELUDE-RULE-CELLS: 26709 -> 211
KMCD-GEN-RULE-CELLS:     27440 -> 205
```