@radumereuta 2022-03-17
KEVM total time spent in front end for proofs (--dry-run)
`radu@radu:~/work/evm-semantics$ time make test-prove -j 8 -O`
Total time dropped from `3m15s` down to `2m56s`

|file | before | after | after/before |
|----|-----------|----|-------|
|address00-spec.k | 20.478s | 16.830s | 0.82|
|ecrecoverloop02-sig0-invalid-spec.k | 20.990s | 16.967s | 0.81|
|ecrecoverloop00-sig0-invalid-spec.k | 21.139s | 17.042s | 0.81|
|bytes00-spec.k | 21.254s | 17.185s | 0.81|
|ecrecoverloop00-sigs-valid-spec.k | 21.182s | 17.221s | 0.81|
|ecrecoverloop02-sig1-invalid-spec.k | 21.218s | 17.449s | 0.82|
|ecrecover00-sigvalid-spec.k | 21.341s | 17.525s | 0.82|
|ecrecover00-siginvalid-spec.k | 21.368s | 17.796s | 0.83|
|ecrecoverloop02-sigs-valid-spec.k | 21.303s | 17.607s | 0.83|
|encodepacked-keccak01-spec.k | 21.149s | 17.526s | 0.83|
|encode-keccak00-spec.k | 21.721s | 17.299s | 0.80|
|overflow00-overflow-spec.k | 21.537s | 17.119s | 0.79|
|functional-spec.k | 16.681s | 17.859s | 1.07|
|requires01-a0gt0-spec.k | 21.745s | 17.752s | 0.82|
|overflow00-nooverflow-spec.k | 21.905s | 17.358s | 0.79|
|keccak00-spec.k | 22.500s | 17.258s | 0.77|
|requires01-a0le0-spec.k | 20.986s | 16.814s | 0.80|
|staticarray00-spec.k | 20.852s | 16.630s | 0.80|
|storagevar01-spec.k | 20.732s | 16.991s | 0.82|
|storagevar00-spec.k | 20.980s | 16.498s | 0.79|
|storagevar02-overflow-spec.k | 20.672s | 16.361s | 0.79|
|storagevar02-nooverflow-spec.k | 20.822s | 16.740s | 0.80|
|staticloop00-a0lt10-spec.k | 21.355s | 17.123s | 0.80|
|storagevar03-spec.k | 21.166s | 17.415s | 0.82|
|lemmas-no-smt-spec.k | 22.552s | 17.021s | 0.75|
|merkle-spec.k | 21.793s | 18.083s | 0.83|
|storageRoot-spec.k | 22.289s | 18.213s | 0.82|
|infinite-gas-spec.k | 23.470s | 18.182s | 0.77|
|allowance-spec.k | 20.226s | 16.115s | 0.80|
|approve-failure-spec.k | 23.024s | 18.429s | 0.80|
|approve-success-spec.k | 23.009s | 19.232s | 0.84|
|lemmas-spec.k | 24.802s | 20.246s | 0.82|
|balanceOf-spec.k | 19.502s | 16.315s | 0.84|
|totalSupply-spec.k | 19.557s | 16.537s | 0.85|
|transfer-failure-1-b-spec.k | 21.202s | 17.127s | 0.81|
|transfer-failure-1-c-spec.k | 20.425s | 16.816s | 0.82|
|transfer-failure-1-a-spec.k | 21.778s | 16.968s | 0.78|
|transfer-failure-2-a-spec.k | 21.047s | 17.597s | 0.84|
|transfer-failure-2-b-spec.k | 20.205s | 16.022s | 0.79|
|transferFrom-failure-1-a-spec.k | 24.273s | 19.432s | 0.80|
|transferFrom-failure-2-b-spec.k | 20.189s | 16.503s | 0.82|
|transferFrom-failure-2-a-spec.k | 21.015s | 16.061s | 0.76|
|transferFrom-failure-1-c-spec.k | 22.867s | 16.632s | 0.73|
|transferFrom-failure-2-c-spec.k | 21.105s | 19.423s | 0.92|
|transferFrom-success-1-spec.k | 22.592s | 16.855s | 0.75|
|transferFrom-failure-1-b-spec.k | 26.393s | 22.608s | 0.86|
|transferFrom-success-2-spec.k | 20.892s | 17.675s | 0.85|
|transferFrom-failure-1-d-spec.k | 28.260s | 23.718s | 0.84|
|transfer-success-2-spec.k | 20.400s | 16.202s | 0.79|
|approve-spec.k | 20.344s | 16.817s | 0.83|
|transfer-success-1-spec.k | 21.402s | 18.322s | 0.86|
|transfer-failure-1-spec.k | 20.878s | 16.301s | 0.78|
|transfer-failure-2-spec.k | 21.224s | 15.564s | 0.73|
|transferFrom-failure-2-spec.k | 20.702s | 16.512s | 0.80|
|transferFrom-failure-1-spec.k | 22.936s | 16.278s | 0.71|
|forwardToHotWallet-failure-1-spec.k | 19.647s | 17.194s | 0.88|
|forwardToHotWallet-failure-2-spec.k | 20.407s | 17.568s | 0.86|