@radumereuta 2022-03-17
C-semantics (make):
Notice the time difference when parsing configurations and rules.
Before:
```
Outer parsing [212 modules]                                  =  0.998s
  New scanner: STDOUT-STREAM-CONFIG-CELLS                    =  0.199s
  New scanner: STDIN-STREAM-CONFIG-CELLS                     =  0.222s
  New scanner: C-CONFIGURATION-CONFIG-CELLS                  =  0.412s
Parse configurations [3/3 declarations]                      =  5.711s
  New scanner: C-RULE-CELLS                                  =  2.325s
  New scanner: STDIN-STREAM-RULE-CELLS                       =  0.543s
  New scanner: STDOUT-STREAM-RULE-CELLS                      =  0.398s
Parse rules [2507/2507 rules]                                = 01m 04s
Validate definition                                          =  0.727s
Apply compile pipeline                                       =  8.565s
Kompile to kore                                              =  2.516s
Save to disk                                                 =  0.760s
Backend                                                      = 59.192s
Cleanup                                                      =  0.001s
Total                                                        = 02m 22s
```
After:
```
Outer parsing [212 modules]                                  =  0.972s
  New scanner: STDOUT-STREAM-CONFIG-CELLS                    =  0.177s
  New scanner: STDIN-STREAM-CONFIG-CELLS                     =  0.206s
  New scanner: C-CONFIGURATION-CONFIG-CELLS                  =  0.398s
Parse configurations [3/3 declarations]                      =  1.251s
  New scanner: C-RULE-CELLS                                  =  2.211s
  New scanner: STDIN-STREAM-RULE-CELLS                       =  0.348s
  New scanner: STDOUT-STREAM-RULE-CELLS                      =  0.418s
Parse rules [2507/2507 rules]                                = 44.869s
Validate definition                                          =  0.730s
Apply compile pipeline                                       =  7.597s
Kompile to kore                                              =  2.446s
Save to disk                                                 =  0.820s
Backend                                                      = 58.603s
Cleanup                                                      =  0.000s
Total                                                        = 01m 57s
```