@Baltoli 2022-03-21
Yeah I'd be interested to see a high-level breakdown of the changes as well - this isn't part of the code that I'm familiar with